update Product controller & Category controller and some views

+++ database/migrations/2019_05_08_234619_create_billings_table.php
@@ -15,7 +15,7 @@ class CreateBillingsTable extends Migration
     {
         Schema::create('billings', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('order_id');
+            $table->unsignedTinyInteger('order_id');
             $table->integer('product_id');
             $table->integer('quantity');
             $table->decimal('product_unit_price',8,2);
